requirements_updated.txt
bcrypt>=4.0.1
flask-cors>=4.0.0
flask>=2.3.3
gurobipy>=10.0.0
matplotlib>=3.7.2
numpy>=1.24.3
pandas>=2.0.3
psycopg2-binary>=2.9.6
pugsql>=0.3.1
pyjwt>=2.6.0
pyomo>=6.6.1
pytest>=7.3.1
python-dotenv>=1.0.0
requests>=2.31.0
schematics>=2.1.1
sentry-sdk[flask]>=1.28.1
sqlalchemy>=2.0.19
tqdm>=4.65.0
us>=3.1.1
xlrd>=2.0.1
openpyxl>=3.1.2  # Modern Excel file support
